+++ src/react/index.ts
@@ -70,6 +70,7 @@ export {
 };
 export type { ReactRootOptions };
 
+export * from "./boot.js";
 export * from "#c55llzkpl4ob";
 export * from "#ft8e49grjdee";
 export * from "#nn6lx6ueg6es";